feat : 인격별 각자 설정
