+++ src/chating/ChatComponents.tsx
@@ -1,14 +1,8 @@
-import React, { useState, useEffect, useRef, useCallback } from 'react';
+import React, { useRef } from 'react';
 import html2canvas from 'html2canvas';
-import { useNavigate, useParams } from 'react-router-dom';
-import TypingWaiting from './component/TypingWaiting';
+import { useNavigate } from 'react-router-dom';
 import { Character } from '../assets/initCharacter';
-import { ReactComponent as Leftarrow } from '../assets/svg/leftarrow.svg';
-import { ReactComponent as Closeicon } from '../assets/svg/closeIcon.svg';
-import {
-  CloseButton, ChatBox, CharacterChat, UserChat, CharacterChatContent, UserChatContent, CharacterAvatar, CharacterMessage, UserMessage, CharacterProfile,
-  ProfileName, UserInputCon, InputMessage, SendButton, MicButton, AlertOverlay, AlertBox, AlertButtons, AlertButtonCancle, AlertButtonFinish,
-} from './component/chatingStyles';
+import { AlertOverlay, AlertBox, AlertButtons, AlertButtonCancle, AlertButtonFinish } from './component/chatingStyles';
 import { GmarketSansMedium } from '../assets/styles';
 import {
   LogNickname, ChatImage, ChatSummary, Solution, PersonalitySection, PersonalityImage, PersonalityDescription, LogDate, LogHeaderImage,
@@ -18,135 +12,6 @@ import penImg from '../assets/png/pen.png';
 import logpersonaImg from '../assets/png/logpersona.png';
 import chatImg from '../assets/png/uncleback.png';
 import downloadImg from '../assets/png/download.png';
-import { debounce } from 'lodash';
-
-export const UPCharacterProfile: React.FC<{ name: string; onClose: () => void; fontFamily?: string }> = ({ name, onClose, fontFamily }) => {
-  const navigate = useNavigate();
-  const { nickname } = useParams<{ nickname: string }>();
-
-  return (
-    <CharacterProfile>
-      <Leftarrow onClick={() => navigate(-1)} />
-      <ProfileName style={{ fontFamily }}>{name}</ProfileName>
-      <CloseButton onClick={onClose}>
-        <Closeicon />
-      </CloseButton>
-    </CharacterProfile>
-  );
-};
-
-interface CharacterChatContentProps {
-  isTyping: boolean;
-  chatEndRef: React.RefObject<HTMLDivElement>;
-  children: React.ReactNode;
-  backgroundColor?: string;
-  fontFamily?: string;
-}
-
-export const CharacterChatCon: React.FC<CharacterChatContentProps> = ({ isTyping, chatEndRef, children, backgroundColor, fontFamily }) => {
-  const [showText, setShowText] = useState<React.ReactNode>(null);
-  const [showTyping, setShowTyping] = useState(true);
-
-  useEffect(() => {
-    if (isTyping) {
-      setShowText(null);
-      setShowTyping(true);
-    } else {
-      const timer = setTimeout(() => {
-        setShowText(children);
-        setShowTyping(false);
-      }, 1000);
-
-      return () => clearTimeout(timer);
-    }
-  }, [isTyping, children]);
-
-  useEffect(() => {
-    if (showText && chatEndRef.current) {
-      chatEndRef.current.scrollIntoView({ behavior: 'smooth', block: 'end' });
-    }
-  }, [showText, chatEndRef]);
-
-  return (
-    <CharacterChatContent style={{ backgroundColor, fontFamily }}>
-      {showTyping ? <TypingWaiting /> : showText}
-    </CharacterChatContent>
-  );
-};
-
-export const ChatingBox: React.FC<{ messages: { text: string; isUser: boolean }[]; isTyping: boolean; character: Character }> = ({ messages, isTyping, character }) => {
-  const chatEndRef = useRef<HTMLDivElement | null>(null);
-
-  useEffect(() => {
-    if (chatEndRef.current) {
-      chatEndRef.current.scrollIntoView({ behavior: 'smooth', block: 'end' });
-    }
-  }, [messages, isTyping]);
-
-  return (
-    <ChatBox>
-      {messages.map((msg, index) => (
-        <div key={index} className={msg.isUser ? UserMessage : CharacterMessage}>
-          {msg.isUser ? (
-            <UserChat>
-              <UserChatContent>{msg.text}</UserChatContent>
-            </UserChat>
-          ) : (
-            <CharacterChat>
-              <CharacterAvatar src={character.img} alt="Character Avatar" />
-              <CharacterChatCon
-                isTyping={index === messages.length - 1 && isTyping}
-                chatEndRef={chatEndRef}
-                backgroundColor={character.background}
-                fontFamily={character.fontFamily}
-              >
-                {msg.text}
-              </CharacterChatCon>
-            </CharacterChat>
-          )}
-        </div>
-      ))}
-      <div ref={chatEndRef} />
-    </ChatBox>
-  );
-};
-
-export const UserInputBox: React.FC<{ input: string; setInput: (input: string) => void; sendMessage: (message: string) => void; handleStartSTT: () => void; handleEndSTT: () => void; }> = ({ input, setInput, sendMessage, handleStartSTT, handleEndSTT }) => {
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  const handleSend = useCallback(debounce(() => {
-    const sanitizedInput = input.replace(/(\r\n|\n|\r)/gm, '').trim(); // 개행 문자 제거 및 불필요한 공백 제거
-    if (sanitizedInput) {
-      sendMessage(sanitizedInput);
-      setInput('');
-      if (inputRef.current) {
-        inputRef.current.value = '';
-      }
-    }
-  }, 300), [input, sendMessage, setInput]);
-
-  const handleKeyDown = useCallback((e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      e.preventDefault();
-      handleSend();
-    }
-  }, [handleSend]);
-
-  return (
-    <UserInputCon>
-      <InputMessage
-        type="text"
-        value={input}
-        onChange={(e) => setInput(e.target.value)}
-        onKeyDown={handleKeyDown}
-        placeholder="메시지를 입력하세요."
-        ref={inputRef}
-      />
-      <SendButton onClick={handleSend}>Send</SendButton>
-      <MicButton onMouseDown={handleStartSTT} onMouseUp={handleEndSTT}>Hold to Speak</MicButton>
-    </UserInputCon>
-  );
-};
 
 export const CustomAlert: React.FC<{ message: string; onConfirm: () => void; onCancel: () => void }> = ({ message, onConfirm, onCancel }) => {
   return (
